feat(websocket): Add WebSocket implementation with Nest.js and Vue.js
This update includes the implementation of a simple client and server WebSocket using Nest.js and Vue.js for a chat application. It is a study project aiming to demonstrate real-time messaging between server (using Nest.js) and client (a Vue.js application). The code both ensures WebSocket connection and handles messaging.